pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "exceland-factory"
version = "0.1.0"
description = "Factoría semiautomática de plantillas Excel comerciales desde specs YAML"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "xlsxwriter>=3.2.0",
    "openpyxl>=3.1.2",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.1",
    "typer>=0.12.3",
    "rich>=13.7.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
exceland = "exceland_factory.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/exceland_factory"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
src = ["src"]
line-length = 100
